Update fork without a local round trip

+++ b2g/components/AlertsService.js
@@ -21,7 +21,6 @@ XPCOMUtils.defineLazyServiceGetter(this, "notificationStorage",
                                    "@mozilla.org/notificationStorage;1",
                                    "nsINotificationStorage");
 
-
 XPCOMUtils.defineLazyGetter(this, "cpmm", function() {
   return Cc["@mozilla.org/childprocessmessagemanager;1"]
            .getService(Ci.nsIMessageSender);
@@ -42,6 +41,7 @@ const kMessageAppNotificationReturn  = "app-notification-return";
 const kMessageAlertNotificationSend  = "alert-notification-send";
 const kMessageAlertNotificationClose = "alert-notification-close";
 
+const kTopicAlertShow          = "alertshow";
 const kTopicAlertFinished      = "alertfinished";
 const kTopicAlertClickCallback = "alertclickcallback";
 
@@ -66,20 +66,36 @@ AlertsService.prototype = {
   },
 
   // nsIAlertsService
+  showAlert: function(aAlert, aAlertListener) {
+    if (!aAlert) {
+      return;
+    }
+    cpmm.sendAsyncMessage(kMessageAlertNotificationSend, {
+      imageURL: aAlert.imageURL,
+      title: aAlert.title,
+      text: aAlert.text,
+      clickable: aAlert.textClickable,
+      cookie: aAlert.cookie,
+      listener: aAlertListener,
+      id: aAlert.name,
+      dir: aAlert.dir,
+      lang: aAlert.lang,
+      dataStr: aAlert.data,
+      inPrivateBrowsing: aAlert.inPrivateBrowsing
+    });
+  },
+
   showAlertNotification: function(aImageUrl, aTitle, aText, aTextClickable,
                                   aCookie, aAlertListener, aName, aBidi,
-                                  aLang) {
-    cpmm.sendAsyncMessage(kMessageAlertNotificationSend, {
-      imageURL: aImageUrl,
-      title: aTitle,
-      text: aText,
-      clickable: aTextClickable,
-      cookie: aCookie,
-      listener: aAlertListener,
-      id: aName,
-      dir: aBidi,
-      lang: aLang
-    });
+                                  aLang, aDataStr, aPrincipal,
+                                  aInPrivateBrowsing) {
+    let alert = Cc["@mozilla.org/alert-notification;1"].
+      createInstance(Ci.nsIAlertNotification);
+
+    alert.init(aName, aImageUrl, aTitle, aText, aTextClickable, aCookie,
+               aBidi, aLang, aDataStr, aPrincipal, aInPrivateBrowsing);
+
+    this.showAlert(alert, aAlertListener);
   },
 
   closeAlert: function(aName) {
@@ -94,6 +110,7 @@ AlertsService.prototype = {
     let uid = (aDetails.id == "") ?
           "app-notif-" + uuidGenerator.generateUUID() : aDetails.id;
 
+    let dataObj = this.deserializeStructuredClone(aDetails.data);
     this._listeners[uid] = {
       observer: aAlertListener,
       title: aTitle,
@@ -105,7 +122,8 @@ AlertsService.prototype = {
       dbId: aDetails.dbId || undefined,
       dir: aDetails.dir || undefined,
       tag: aDetails.tag || undefined,
-      timestamp: aDetails.timestamp || undefined
+      timestamp: aDetails.timestamp || undefined,
+      dataObj: dataObj || undefined
     };
 
     cpmm.sendAsyncMessage(kMessageAppNotificationSend, {
@@ -136,31 +154,60 @@ AlertsService.prototype = {
       // notification via a system message containing the title/text/icon of
       // the notification so the app get a change to react.
       if (data.target) {
-        gSystemMessenger.sendMessage(kNotificationSystemMessageName, {
-            clicked: (topic === kTopicAlertClickCallback),
-            title: listener.title,
-            body: listener.text,
-            imageURL: listener.imageURL,
-            lang: listener.lang,
-            dir: listener.dir,
-            id: listener.id,
-            tag: listener.tag,
-            dbId: listener.dbId,
-            timestamp: listener.timestamp
-          },
-          Services.io.newURI(data.target, null, null),
-          Services.io.newURI(listener.manifestURL, null, null)
-        );
+        if (topic !== kTopicAlertShow) {
+          // excluding the 'show' event: there is no reason a unlaunched app
+          // would want to be notified that a notification is shown. This
+          // happens when a notification is still displayed at reboot time.
+          gSystemMessenger.sendMessage(kNotificationSystemMessageName, {
+              clicked: (topic === kTopicAlertClickCallback),
+              title: listener.title,
+              body: listener.text,
+              imageURL: listener.imageURL,
+              lang: listener.lang,
+              dir: listener.dir,
+              id: listener.id,
+              tag: listener.tag,
+              timestamp: listener.timestamp,
+              data: listener.dataObj || undefined,
+            },
+            Services.io.newURI(data.target, null, null),
+            Services.io.newURI(listener.manifestURL, null, null)
+          );
+        }
+      }
+      if (topic === kTopicAlertFinished && listener.dbId) {
+        notificationStorage.delete(listener.manifestURL, listener.dbId);
       }
     }
 
     // we're done with this notification
     if (topic === kTopicAlertFinished) {
-      if (listener.dbId) {
-        notificationStorage.delete(listener.manifestURL, listener.dbId);
-      }
       delete this._listeners[data.uid];
     }
+  },
+
+  deserializeStructuredClone: function(dataString) {
+    if (!dataString) {
+      return null;
+    }
+    let scContainer = Cc["@mozilla.org/docshell/structured-clone-container;1"].
+      createInstance(Ci.nsIStructuredCloneContainer);
+
+    // The maximum supported structured-clone serialization format version
+    // as defined in "js/public/StructuredClone.h"
+    let JS_STRUCTURED_CLONE_VERSION = 4;
+    scContainer.initFromBase64(dataString, JS_STRUCTURED_CLONE_VERSION);
+    let dataObj = scContainer.deserializeToVariant();
+
+    // We have to check whether dataObj contains DOM objects (supported by
+    // nsIStructuredCloneContainer, but not by Cu.cloneInto), e.g. ImageData.
+    // After the structured clone callback systems will be unified, we'll not
+    // have to perform this check anymore.
+    try {
+      let data = Cu.cloneInto(dataObj, {});
+    } catch(e) { dataObj = null; }
+
+    return dataObj;
   }
 };
