+++ ipc/chromium/src/base/process_util_posix.cc
@@ -22,11 +22,12 @@
 #include "base/logging.h"
 #include "base/platform_thread.h"
 #include "base/process_util.h"
-#include "base/scoped_ptr.h"
 #include "base/sys_info.h"
 #include "base/time.h"
 #include "base/waitable_event.h"
 #include "base/dir_reader_posix.h"
+
+#include "mozilla/UniquePtr.h"
 
 const int kMicrosecondsPerSecond = 1000000;
 
@@ -47,9 +48,14 @@ bool OpenProcessHandle(ProcessId pid, ProcessHandle* handle) {
   return true;
 }
 
-bool OpenPrivilegedProcessHandle(ProcessId pid, ProcessHandle* handle) {
+bool OpenPrivilegedProcessHandle(ProcessId pid,
+                                 ProcessHandle* handle,
+                                 int64_t* error) {
   // On POSIX permissions are checked for each operation on process,
   // not when opening a "handle".
+  if (error) {
+    *error = 0;
+  }
   return OpenProcessHandle(pid, handle);
 }
 
@@ -70,16 +76,21 @@ bool KillProcess(ProcessHandle process_id, int exit_code, bool wait) {
 
   if (result && wait) {
     int tries = 60;
+    bool exited = false;
     // The process may not end immediately due to pending I/O
     while (tries-- > 0) {
       int pid = HANDLE_EINTR(waitpid(process_id, NULL, WNOHANG));
-      if (pid == process_id)
+      if (pid == process_id) {
+        exited = true;
         break;
+      }
 
       sleep(1);
     }
 
-    result = kill(process_id, SIGKILL) == 0;
+    if (!exited) {
+      result = kill(process_id, SIGKILL) == 0;
+    }
   }
 
   if (!result)
@@ -101,7 +112,7 @@ class ScopedDIRClose {
     }
   }
 };
-typedef scoped_ptr_malloc<DIR, ScopedDIRClose> ScopedDIR;
+typedef mozilla::UniquePtr<DIR, ScopedDIRClose> ScopedDIR;
 
 
 void CloseSuperfluousFds(const base::InjectiveMultimap& saved_mapping) {
